1.fix some mistake 2.增加消息删除, 阅读记录删除接口

+++ hami-security/src/main/java/top/wang3/hami/security/ratelimit/handler/FixedWindowRateLimiterHandler.java
@@ -28,7 +28,7 @@ public class FixedWindowRateLimiterHandler implements RateLimiterHandler {
 
     @Override
     public List<Long> execute(String key, double rate, double capacity) {
-        double interval = capacity / rate;
+        long interval =  (long) (capacity / rate);
         return RedisClient.executeScript(fixedWindowScript, List.of(key), List.of(interval, capacity));
     }
 }
